+++ .github/assets/testing/edge.yml
@@ -54,11 +54,11 @@ features:
     software:
       charms:
         ironic-conductor-k8s:
-          channel: 2025.1/beta
+          channel: 2025.1/edge
         ironic-k8s:
-          channel: 2025.1/beta
+          channel: 2025.1/edge
         nova-ironic-k8s:
-          channel: 2025.1/beta
+          channel: 2025.1/edge
   caas:
     software:
       charms:
