+++ sunbeam-python/sunbeam/features/baremetal/constants.py
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: 2025 - Canonical Ltd
+# SPDX-License-Identifier: Apache-2.0
+
+IRONIC_CONDUCTOR_APP = "ironic-conductor"
+IRONIC_APP_TIMEOUT = 600
