+++ sunbeam-python/sunbeam/features/baremetal/feature.py
@@ -2,27 +2,18 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import logging
-import queue
 
 import click
 from packaging.version import Version
 from rich.console import Console
-from rich.status import Status
 
 from sunbeam.core.common import (
     BaseStep,
-    Result,
-    ResultType,
     run_plan,
-    update_status_background,
 )
 from sunbeam.core.deployment import Deployment
 from sunbeam.core.juju import (
-    ActionFailedException,
     JujuHelper,
-    JujuStepHelper,
-    JujuWaitException,
-    LeaderNotFoundException,
 )
 from sunbeam.core.manifest import (
     AddManifestStep,
@@ -30,10 +21,10 @@ from sunbeam.core.manifest import (
     FeatureConfig,
     SoftwareConfig,
 )
-from sunbeam.core.openstack import OPENSTACK_MODEL
 from sunbeam.core.terraform import (
     TerraformInitStep,
 )
+from sunbeam.features.baremetal import steps
 from sunbeam.features.interface.v1.openstack import (
     EnableOpenStackApplicationStep,
     OpenStackControlPlaneFeature,
@@ -42,62 +33,8 @@ from sunbeam.features.interface.v1.openstack import (
 from sunbeam.utils import click_option_show_hints, pass_method_obj
 from sunbeam.versions import OPENSTACK_CHANNEL
 
-IRONIC_CONDUCTOR_APP = "ironic-conductor"
-IRONIC_APP_TIMEOUT = 600
 LOG = logging.getLogger(__name__)
 console = Console()
-
-
-class RunSetTempUrlSecretStep(BaseStep, JujuStepHelper):
-    """Run the set-temp-url-secret action on the ironic-conductor."""
-
-    def __init__(
-        self,
-        deployment: Deployment,
-        jhelper: JujuHelper,
-    ):
-        super().__init__(
-            "Run the set-temp-url-secret action on ironic-conductor",
-            "Running the set-temp-url-secret action on ironic-conductor",
-        )
-        self.jhelper = jhelper
-        self.deployment = deployment
-        self.model = OPENSTACK_MODEL
-
-    def run(self, status: Status | None = None) -> Result:
-        """Run the set-temp-url-secret action on ironic-conductor."""
-        try:
-            unit = self.jhelper.get_leader_unit(IRONIC_CONDUCTOR_APP, self.model)
-            self.jhelper.run_action(
-                unit,
-                self.model,
-                "set-temp-url-secret",
-            )
-        except (ActionFailedException, LeaderNotFoundException) as e:
-            LOG.error(
-                "Error running the set-temp-url-secret action on ironic-conductor: %s",
-                e,
-            )
-            return Result(ResultType.FAILED, str(e))
-
-        apps = [IRONIC_CONDUCTOR_APP]
-        LOG.debug(f"Application monitored for readiness: {apps}")
-        status_queue: queue.Queue[str] = queue.Queue()
-        task = update_status_background(self, apps, status_queue, status)
-        try:
-            self.jhelper.wait_until_active(
-                self.model,
-                apps,
-                timeout=IRONIC_APP_TIMEOUT,
-                queue=status_queue,
-            )
-        except (JujuWaitException, TimeoutError) as e:
-            LOG.warning(str(e))
-            return Result(ResultType.FAILED, str(e))
-        finally:
-            task.stop()
-
-        return Result(ResultType.COMPLETED)
 
 
 class BaremetalFeature(OpenStackControlPlaneFeature):
@@ -180,7 +117,7 @@ class BaremetalFeature(OpenStackControlPlaneFeature):
                     # until we run the set-temp-url-secret action.
                     app_desired_status=["active", "blocked"],
                 ),
-                RunSetTempUrlSecretStep(
+                steps.RunSetTempUrlSecretStep(
                     deployment,
                     jhelper,
                 ),
